RSPY-488: Splitted files in rs_server_staging
